requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
click==8.2.1
colorama==0.4.6
fastapi==0.115.12
frozenlist==1.7.0
h11==0.16.0
idna==3.10
multidict==6.4.4
propcache==0.3.2
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
yarl==1.20.1
